+++ libs/engine/tests/framework/test_utilities.py
@@ -146,7 +146,7 @@ def test_handle_exceptions(test_input: type[BaseException]) -> None:
     with pytest.raises(test_input):
         # known issue with mypy
         # see heated thread at https://github.com/python/mypy/issues/6549
-        func = handle_exceptions(raise_me(test_input), None, False)  # type: ignore[func-returns-value]
+        func = handle_exceptions(raise_me(test_input), None, False)  # type: ignore[arg-type, func-returns-value]
         func()